requirements.txt
biopython>=1.79
matplotlib>=3.5.0
numpy>=1.23.0
flask>=3.1.1
gunicorn>=20.1.0
pytest>=7.0.0
